Use getPrefixedURL and getPrefixedDBkey instead of getText
The getText() method of a Title strips out the namespace prefix.
If we want to use the titles in change-prop, the easiest way is
to use titles in the prefixed db-key format. In that case we will
be able to 'just use' the field without additional processing.

Also, the titles which are used in the URIs for resource_change
event should be uri-encoded.

Change-Id: I132d0b7e70e1a992f0204361a71b715c2cd1655f