Bug fix: determine correctly if a revision was made by a bot
Ic79690da726698ee785e048cbb2959c731bd017d introduces the rev_by_bot flag
for revision_create events. However, it was wrongfully calculated and
always yielded 'false' because $flags is actually a string comprised of
a list of comma-separated values. The correct way to see if an edit was
made by a bot is to load the User object and check its bot status via
isBot().

Change-Id: Ibcda64abeee2038905dac8f0ae999eed05ed3a9d