Use LinksUpdate member for page_id
The onLinksUpdateComplete hook is run asyncronously
by the JobQueue, so the article might've been deleted
by the time the hook was run. In that case Title->getArticleID()
might return zero which is not conforming to our event schema.

The mId field, however, should still be correct since it's filled
when the LinksUpdate object is created.

Bug: T158702
Change-Id: Ib32f837937c0c11df60a996ea4d72cf81461906c