Emit mediawiki.page-create event on PageContentInsertComplete
This re-uses the mediawiki/revision/create schema.  As such, I've factored out
the code that creates the revision create schema attributes into a static
function in EventBus.php.

Along the way I noticed that we don't set rev_parent_id if it is == 0.
The rev_parent_id minimum of 1 was removed and set to 0 in e915290c.
So, this patch will now set rev_parent_id = 0 if it is set to that in the
Revision object.

Bug: T150369
Change-Id: Ib8dc8c9957c806681d27a6f4cf5b8b812260513a