Create a lightweight JobRunner that just executes a job.
In the current JobQueue architecture the JobRunner is
responsible for both dequeueing and executing the jobs.
When moving to kafka-based implementation, ChangeProp
will handle the dequeueing part and post the jobs for
execution, so we need a lighter-weight version of the
JobRunner that just executes the job.

The JobExecutor is a bare-bone executor without any
stats reporting, transaction/replica lag management
or other features from the JobRunner. These features
will be added on demand. In future the JobExecutor
will be transfered to the core and the JobRunner will
delegate the actual execution to it, so all the code
duplication will be removed. However right now we need
to set up the whole pipeline to test whether the new
architecture works as a whole and see whether we need
any additional features from the JobRunner at all.

Change-Id: Ib1aaa58a730b20ee266a530150d684ae9782e6d2