Support per-event-type EventBus enabling configuration.
In order to enable the kafka-based job queue for private wikis
we need to first enable EventBus service for private wikis.

However, for now we don't want to post normal mediawiki events
for private wikis because ChangeProp/RESTBase stack doesn't support
that and secondly because the mediawiki events are replicated
cross DC using MirrorMaker that currently doesn't use TLS and we
do not want unencrypted private traffic to flow across DC boundaries.

With this feature we would be able to selectively send certain event
types per wiki with a configuration like the following:

'wgEnableEventBus' => [
    'default' => 'TYPE_ALL',
    'wikitech' => 'TYPE_JOB',
    'private' => 'TYPE_JOB',
    'loginwiki' => 'TYPE_JOB',
    'votewiki' => 'TYPE_JOB,
],

Bug: T191464
Change-Id: Id1d043e5ce02e73b51f75ee54575647b591437f9