Validate that events are JSON-serializble and log and error if not.
In the old queue certain jobs relied on PHP native serialization
and did not reinstantiated classes in the parameters upon deserialization.
The new queue does not support that feature and all the jobs must be
JSON-serializable, thus contain only scalar values in the parameters.

In order to get a full list of jobs that should be modified we need
to check for non-scalar values and log an error if we encounter one.

Bug: T190327
Change-Id: Ib6d341837dfef1986a9aa9fd543ad3371f17361d