Use 'array_values' to renumber the array after 'array_unique'
The 'array_unique', if not a no-op, removes the non-unique elements
from the array, but preserve the items numbering, creatin an array
with holes. such an arrays can't be JSON-encoded into JSON array,
so it's encoded into a JSON object, thus an event fails validation.

To fix that we can use array_values to reorder the tags and fill
the gaps.

Change-Id: If5f2fc95ae80fbeb873a62d5060203d7fb151ccb