[EventBus] Decrese the maximum log size to 1kb
With syslog the max size of the message is 1kb, so for large
events a lot of the messages end up in jsonTruncated channel.

In EventBus we have a way to make messages smaller by only leaving
the .meta part of the event, so do that if the payload is too large
for the log.

Bug: T204183
Change-Id: Id4cf533cd06e5d78f5c8effecc0e3f5110962dbd