Remove unrelated stream name from JobQueue backend exception
This needlessly fragmented the problem, making it harder to detect
as regression when they are all reported separately.

Ideally the HTTP backend error would be included since that is
generally related and can be grouped together. However since a
recent refactor that is no longer possible since send() may no
longer throw directly it seems, and may return an array of errors.

Details of the HTTP error remain logged by EventBus.php as
diagnostic message. Details of the job are now also logged as
diagnostic message.

Bug: T249745
Change-Id: Ie0faec5522b9979d1222ff99698e699df0cce63b