Support x_client_ip_forwarding_enabled setting per event service
Certain analytics schemas require the requesting IP address (as
http.client_ip). Currently, for server-side events, this is being
produced as 127.0.0.1, as requests from MediaWiki to eventgate are
cluster-internal, and no other IP information is provided.

The X-Client-IP header is present on requests on Wikimedia's
production MediaWiki hosts, and is used to propagate the requesting IP
address. If this header is sent along to eventgate with event
production requests, it will be used by eventgate when populating
http.client_ip (if the latter is present in the event schema).

This change updates the event service configurations in the
EventServices config setting to support a new property,
x_client_ip_forwarding_enabled. If x_client_ip_forwarding_enabled is
true, and if the X-Client-IP header is present, the header will be
forwarded to the event intake service.

Bug: T288853
Change-Id: I10dd7fe30dee3a6ce32af4299ffe8c1881638f82