Add ability to disable a stream in EventStreamConfig
This adds an producers.mediawiki_eventbus.enabled
setting to each stream, defaulting to true.

New behavior:
If EventStreamConfig is enabled, and a stream's config
has producers.mediawiki_eventbus.enabled == false,
or if that stream does not have an entry in
EventStreamConfig wgEventStreams, then it is considered
to be disabled. In this case a singleton no-op
EventBus instance will be chosen for 'producing'
this stream, resulting in no events being sent.

This will allow developers to disable or enable
streams produced by EventBus per wiki in mediawiki-config.

Bug: T259712

Change-Id: Ia45939db3778d748bddb27c396e513d09e293ef6