Encode redirect targets in page change events.
Motivation: The CirrusSearch Streaming Updater (flink) is based on the page-change event stream.
CirrusSearch does not index redirects directly but rather updates the
redirect target to keep a record of the redirect. In case of a page
delete event, there is no way to restore the redirect target from
outside the wiki issuing the event, at least not for an anonymous user
which the flink application is. Therefore, we'd like to make the
page-change events more self contained by adding information about the
redirect target of a page.

Bug: T325315
Depends-On: I5f6fe0db3db3a770c2b4d84af5d9bb2491dd530f
Change-Id: I238061d64a3809fac8b66c401253fc6357dd8879