Resolve redirect target before delete, so it can be serialized in the event fired after.
Motivation: The CirrusSearch Streaming Updater (flink) is based on the page-change event stream.
CirrusSearch does not index redirects directly but rather updates the
redirect target to keep a record of the redirect. In case of a page
delete event, there is no way to restore the redirect target from
outside the wiki issuing the event, at least not for an anonymous user
which the flink application is. Therefore, we'd like to make the
page-change events more self contained by adding information about the
redirect target of a page.

Bug: T325315
Change-Id: I31a2548974098274fd866383caf357383f99e876