Filter out null edits
In the referenced Phabricator task, we found a lot of drift between the
time the event was processed (meta.dt) and the time the event was
supposed to have happened (revision.rev_dt).  In some cases, like a page
restore that's inserting old revisions, we expect a big difference in
these timestamps.  But in other cases, we find a difference where there
was no good explanation.  At least one of these seem to be null edits,
so we are filtering them out and looking again once they're no longer
produced.

Bug: T341134
Change-Id: I56c3643638080615e84bf53dcce0ef664f641dab