Move event must contain the destination page title
When running with a primary/replica configuration most data loaded in
the hooks comes from replica databases. This is acceptable in most
cases, such as loading the old revision, but breaks down when trying
to read information about the edit that triggered creating the event.

In the case of page move events it was possible to receive a WikiPage
representing the old revision, which resulted in the pre-move title
being serialized into the event. Correct by requireing that
information to use READ_LATEST.

Bug: T350299
Change-Id: Ia9ebda0641f885f8b4c0cf888fd651cd6996364d