Correct redirect serialization in page change events
When running databases in a primary/replica configuration and changing
a page from regular wikitext, into a redirect, and back again through
edits to the wikitext, the page change event does not always
accurately represent the redirect. Often times it instead encodes the
redirect information that was true prior to the edit, as it loaded
the information from the replica before it received the edit.

In the dependent patch mw core is updated to ensure that
WikiPage::getRedirectTitle returns the correct value after an edit,
and this patch updates the hooks to use that value if available.

Bug: T350299
Depends-On: I3dd75e0acf936aa900e442fb3ff4de9fb3150359
Change-Id: Ifdb634109193a30f5786bc7708d46234fc97449d