Remove support for $wgJobSerialCommitThreshold
This config that has been set to false in production since 2018 (Ie4ea1dc0d3927)
is about to be removed from core without deprecation.

That removal $dbwSerial uncoditionally becomes false which in
turned the whole method into a one-liner and since it was used only
once, we just replaced that.

Bug: T325389
Change-Id: I24f33a15d214d49bfb6c6013e7fd64c7d0eb0086