Add performer fallback to onChangeTagsAfterUpdateTags
If no user is explicitly supplied but we have a RecentChange object,
fall back to reporting the change's performer as the user who added the
tags.

This is probably semantically correct, but may not be.

revision-tags-change is a relatively unmaintaned/deprecated
event stream.  This should be a good fix (better than nothing),
but knowing for sure would require more time than we have to spend
on this right now.

A future entity state change based model for revision tags
should replace this event.

Bug: T352899
Change-Id: I37e1e288f635ddffad9f8d3db7551db58812a9cd