eventbus: add instrumentation to send() method.
This patch adds initial instrumetnation for
events posted to an intake service (EventGate).

We report count metrics for the number of requests
sent, and the total number of events POSTed.

We report count metrics for resposnes, broken down by
status code.

Whenever possible, we include the target stream name
in the metrics label.

Bug: T363587
Change-Id: I8993c51f8bcf447851a61bcc4245efafa235e663