Forward info about original request
Per T297613#7577140, the X-Client-IP header might be undefined in
MediaWiki on k8s. In this instance, EventBus doesn't set the X-Client-IP
header of the request to the event intake service and continues
processing. The event intake service doesn't have enough information to
determine the client IP and so sets the http.client_ip property of the
corresponding event to 127.0.0.1.

Per T297613#7577140 again (with confirmation from SRE), use
WebRequest#getIP() to determine the client IP.

Bug: T368495
Change-Id: I23fc2a935148be04eae748e2442794d5223aa4b2