EventBus: label metrics with event type name.
Label metrics by event type name. If the lookup fails,
fall back to labeling with the $type id parameter. Unknown or invalid ids
will be reported by the `events_are_not_enqueable` metric, which
fires when an event type does not belong to this EventBus instance allow list.

It should not be possible to supply a $type that does not belong
to EVENT_TYPE_NAME. Falling back to a numerical id is just a guard
to help us spot eventual bugs.

Bug: T363587
Change-Id: I2a4870a682c282afd2104d6d989c88a179fda4e4