Move instrumentation feature flag to extension.json
I assume there’s no good reason not to do this other than that the
linked documentation still documents the old way to declare config
variables because it’s (IMHO) outdated; for the recommended way that’s
partly introduced here, see [1].

Ideally, we’d also inject a Config (or even ServiceOptions) into the
class instead of accessing the global, like we do in e.g. EventFactory
or RunSingleJobHandler, but I think that’s less urgent (especially if
this feature flag is temporary).

[1]: https://www.mediawiki.org/wiki/Manual:Configuration_for_developers#Configuration_using_extension.json_(recommended)

Bug: T363587
Change-Id: Ic94655d400439a1beedb5e3e98a7dd71fd264783