EventBus send - Fix json_decode bug in error metric handling
When we get a failure response from eventgate, the body includes the
events that failed parsing.  We parse these to get the stream name
for metrics.  Previous code expected the failed events to be returned
as JSON encoded strings; eventgate actually returns them in the main
JSON response body.  They do not need to be decoded again.

Bug: T371432
Bug: T363587
Change-Id: I7cc262735bac839087509edeebd722c9088a143b