EventBus - DRY code and rename some metrics and standardize labels
1. DRY code to avoid differing behavior for string $events.

send accepts $events as a string as well as an assoc array of $events.
This was done to avoid extra serialization in cases where the caller
has already serialized events to a JSON string.

However, we always need the $events assoc array in order to emit useful metrics.
In the future, we should deprecate the string support and only accept assoc array.
The number of callers that pass a string is probably small.

For now, this change just DRYs the code so that if $events is passed as a string,
we deserialize it back to an array earlier in the function.

2. rename some metrics and standardize labels

This makes the metric naming a little more consistent:
- event_batch_* is about the batch of $events given to the send method
- events_(outgoing|accepted|failed)_total is about number of events
- events_*_by_stream_total is about number of events with stream_name label
  (note: we cannot easily get events_accepted_by_stream_total, so we don't.)
- event_service_response_total is about http responses from event intake service

Bug: T363587
Change-Id: Ia3faec55f2031809fcec75331bb2eddfddac9050