metrics: provide a fallback value for emtpy labels
When a metric label is initialized but has no value assigned,
fallback to a default string.

Add logging to capure this scenario on prod deployments.

This change refactors a constant previously used to fallback
and label missing stream names, to its more generic case.
The default __stream_unknwon__ value has been renamed to
__value_unknown__.

Bug: T373086
Change-Id: Id9ee80d1573db99bf2673be6f047574c682c6435