PageEntitySerialzer: avoid using WikiPage::isRedirect()
We can't rely on WikiPage instances to test for redirection.

The WikiPage::isRedriect() method will perform a database lookup
to retrieve LinkTarget info.

When handling serialization for deletion events (either from
hooks or domain events) the method will always return false,
because the lookup happens post transaction, and the
redirect info is in not available in the database anymore.

Bug: T393757
Change-Id: Ic946d62c133f07f17197bafddd57dd381b5ec517