EventSerializer - fix logic for setting of meta.dt
When EventSerializer was created in 197b2de4,
The documentation of the ingestionTimestamp parameter
did not match the logic!

This means that the logic implemented did not match the
convention decided on in T267648.

This change fixes the logic so that if ingestionTimestamp
is not provided, meta.dt will not be set in the event.

Bug: T267648
Bug: T376026
Depends-On: Ic67b7dbe5ee29a866a504ecdfe8d16a184e39d1d
Change-Id: Ie8c078017412b3fdcab8b4d7d7bd220ddb08c595